+++ app/Livewire/MemberFilterTrait.php
@@ -3,15 +3,22 @@
 namespace App\Livewire;
 
 use App\Models\MemberFilter;
+use App\Models\MemberGroup;
 
 trait MemberFilterTrait
 {
+    public string $filterMemberGroup = '';
     public bool $filterShowBeforeEntrance = false;
     public bool $filterShowAfterRetired = false;
     public bool $filterShowPaused = false;
 
     public function getMemberFilter(): MemberFilter
     {
-        return new MemberFilter($this->filterShowBeforeEntrance, $this->filterShowAfterRetired, $this->filterShowPaused);
+        $selectedMemberGroup = MemberGroup::query()->find(intval($this->filterMemberGroup));
+        return new MemberFilter(
+            $this->filterShowBeforeEntrance,
+            $this->filterShowAfterRetired,
+            $this->filterShowPaused,
+            $selectedMemberGroup);
     }
 }