+++ app/Livewire/Members/MemberOverview.php
@@ -12,22 +12,10 @@ class MemberOverview extends Component
 {
     use MemberFilterTrait;
 
-    public string $filterMemberGroup = '';
 
     public function getMembersProperty()
     {
-
-        /** @var Builder $memberList */
         $memberList = Member::getAllFiltered($this->getMemberFilter());
-
-        if ($this->filterMemberGroup) {
-            /** @var MemberGroup $selectedGroup */
-            $selectedGroup = MemberGroup::query()->find(intval($this->filterMemberGroup));
-            $groupChildList = $selectedGroup?->getAllChildrenRecursive() ?: [];
-            $memberList->whereHas('memberGroups', function ($query) use ($groupChildList) {
-                $query->whereIn('id', array_map(fn ($group) => $group->id, $groupChildList));
-            });
-        }
 
         return $memberList;
     }