+++ app/Models/MemberFilter.php
@@ -8,17 +8,24 @@ class MemberFilter
     public bool $inclAfterRetired;
     public bool $inclPaused;
 
+    public ?MemberGroup $memberGroup;
 
     /**
      * @param bool $inclBeforeEntrance
      * @param bool $inclAfterRetired
      * @param bool $inclPaused
+     * @param MemberGroup|null $memberGroup
      */
-    public function __construct(bool $inclBeforeEntrance = false, bool $inclAfterRetired = false, bool $inclPaused = false)
+    public function __construct(
+        bool $inclBeforeEntrance = false,
+        bool $inclAfterRetired = false,
+        bool $inclPaused = false,
+        ?MemberGroup $memberGroup = null)
     {
         $this->inclBeforeEntrance = $inclBeforeEntrance;
         $this->inclAfterRetired = $inclAfterRetired;
         $this->inclPaused = $inclPaused;
+        $this->memberGroup = $memberGroup;
     }